Move the websocket server code in to its own module
